% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateCV.R
\name{evaluateCV}
\alias{evaluateCV}
\title{Performance evaluation by cross-validation}
\usage{
evaluateCV(mypredictor = c("predictorKMR", "predictorElasticNet",
  "predictorLasso", "predictorRF"), celllines, celllinesKernel, chemicals,
  chemicalsKernel, toxicity, nfolds = 5, nrepeats = 10, seed = 47,
  mc.cores = 1, ...)
}
\arguments{
\item{mypredictor}{Character indicating a predictor function. Possible options are KMR (default), ElasticNet, Lasso and RF.}

\item{celllines}{Matrix of descriptors for \code{ncell} cell lines, of dimension \code{ncell x pcell}. Only used for ElasticNet, Lasso and RF.}

\item{celllinesKernel}{Kernel Gram matrix for \code{ncell} cell lines, of dimension \code{ncell x ncell}. Only used for KMR.}

\item{chemicals}{Matrix of descriptors for \code{nchem} chemicals, of dimension \code{nchem x pchem}. Only used for ElasticNet, Lasso and RF.}

\item{chemicalsKernel}{Kernel Gram matrix for \code{nchem} chemicals, of dimension \code{nchem x nchem}. Only used for KMR.}

\item{toxicity}{Matrix of toxicity values for \code{ncell} cell lines responding to \code{nchem} chemicals, of dimension \code{ncell x nchem}.}

\item{nfolds}{Number of folds for cross-validation. Default is 5.}

\item{nrepeats}{Number of times the k-fold cross-validation is performed. Default is 1.}

\item{seed}{A seed number for the random number generator (useful to have the same CV splits).}

\item{mc.cores}{Number of parallelable CPU cores to use.}

\item{...}{Other arguments to pass to predictor function.}
}
\value{
A list with matrices of cross-validation performance scores.
Each score matrix is of dimension \code{nexp x nchem} (per CV experiment, per chemical) where \code{nexp=nfolds*nrepeats}
and corresponds to one of the evaluation criteria:
\item{matrix.ci}{Concordance index.}
\item{matrix.rho}{Pearson correlation.}
}
\description{
Evaluate regression performance of a predictor by cross-validation.
}
\references{
Bernard, E., Jiao, Y., Scornet, E., Stoven, V., Walter, T., and Vert, J.-P. (2017). "Kernel multitask regression for toxicogenetics." \href{https://doi.org/10.1101/171298}{bioRxiv-171298}.
}
\seealso{
\code{\link{predictorKMR}}, \code{\link{predictorElasticNet}}, \code{\link{predictorLasso}}, \code{\link{predictorRF}}
}
