% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nontoxic.R
\name{nontoxic}
\alias{nontoxic}
\title{Non-toxic chemical compounds}
\usage{
nontoxic()
}
\value{
A sequence of identifiers for those 15 non-toxic chemical compounds in toxicity data.
}
\description{
15 out of 106 chemical compounds in toxicity data that were shown to have no toxicity across the human cell population.
}
\references{
Eduati, F., et al. "Prediction of human population responses to toxic compounds by a collaborative competition." Nature biotechnology 33.9 (2015): 933-940. \href{https://doi.org/10.1038/nbt.3299}{doi:10.1038/nbt.3299}.
}
