% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictorKMR.R
\name{predictorKMR}
\alias{predictorKMR}
\title{Wrapper function for kernel multitask regression}
\usage{
predictorKMR(patientsKernelTrain, patientsKernelTest, response, drugsKernel,
  lambdas = exp(-15:25), nfolds = 5, nrepeats = 1)
}
\arguments{
\item{patientsKernelTrain}{Precomputed kernel Gram matrix of \code{n} training patients, of dimension \code{n x n}.}

\item{patientsKernelTest}{Precomputed kernel Gram matrix of \code{m} test patients crossing \code{n} training patients, of dimension \code{m x n}.}

\item{response}{Matrix of observed toxicity values, of dimension \code{n x t}, for the \code{n} training patients responding to \code{t} drugs.}

\item{drugsKernel}{Kernel Gram matrix of the \code{t} drugs, of dimension \code{t x t}.}

\item{lambdas}{Sequence of lambdas that must be tested to fit a cross-validated KMR model. Default is exp(-15:25).}

\item{nfolds}{Number of folds for cross-validation. Default is 5.}

\item{nrepeats}{Number of times the k-fold cross-validation is performed. Default is 1.}
}
\value{
A matrix of predicted toxicity values, of dimension \code{m x t}, for the \code{m} test patients responding to the \code{t} drugs.
}
\description{
Wrapper function to perform kernel multitask regression with \code{cv.kmr} that trains a model on training set and then predicts on test set for multiple tasks.
}
\note{
Multitask prediction is made, for which task relationships are encoded in \code{drugsKernel}.
}
\references{
Bernard, E., Jiao, Y., Scornet, E., Stoven, V., Walter, T., and Vert, J.-P. (2017). "Kernel multitask regression for toxicogenetics." \href{https://doi.org/10.1101/171298}{bioRxiv-171298}.
}
\seealso{
\code{\link[kmr]{cv.kmr}}
}
