% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictorRF.R
\name{predictorRF}
\alias{predictorRF}
\title{Wrapper function for random forest regression}
\usage{
predictorRF(patientsTrain, patientsTest, response, ntree = 500)
}
\arguments{
\item{patientsTrain}{Matrix of training descriptors, of dimension \code{n x p}, for \code{n} training patients with \code{p} descriptors.}

\item{patientsTest}{Matrix of test descriptors, of dimension \code{m x p}, for \code{m} test patients with the same set of \code{p} descriptors.}

\item{response}{Matrix of observed toxicity values, of dimension \code{n x t}, for the \code{n} training patients responding to \code{t} drugs.}

\item{ntree}{Number of trees to grow a random forest. Default is 500. All other arguments are taken by default implementation of \code{randomForest}.}
}
\value{
A matrix of predicted toxicity values, of dimension \code{m x t}, for the \code{m} test patients responding to the \code{t} drugs.
}
\description{
Wrapper function to perform random forest regression with \code{randomForest} that trains a model on training set and then predicts on test set for multiple tasks.
}
\note{
Prediction is made per task with no special treatment for multitask learning, nor are task features needed.
}
\seealso{
\code{\link[randomForest]{randomForest}}
}
