% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tox.R
\docType{data}
\name{tox}
\alias{tox}
\title{The toxicity matrix}
\format{A data frame of EC10 values of 884 cell lines in rows and 91 toxic chemical compounds in columns.}
\source{
\url{https://doi.org/10.7303/syn1761567}
}
\usage{
tox
}
\description{
The toxicity values of cell lines, represented by the one-tenth maximal effective concentration-response exposure data (EC10) responding to various chemicals, from the Dream 8 toxicogenetics challenge dataset.
}
\keyword{datasets}
