% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.rna.R
\docType{data}
\name{id.rna}
\alias{id.rna}
\title{The identifiers of RNA-seq only cell lines}
\format{A vector of character strings of identifiers of the 337 RNA-seq only cell lines.}
\source{
\url{https://doi.org/10.7303/syn1761567}
}
\usage{
id.rna
}
\description{
The identifiers of RNA-seq only cell lines of the Dream 8 toxicogenetics challenge dataset.
}
\references{
Eduati, F., et al. "Prediction of human population responses to toxic compounds by a collaborative competition." Nature biotechnology 33.9 (2015): 933-940. \href{https://doi.org/10.1038/nbt.3299}{doi:10.1038/nbt.3299}.
}
\keyword{datasets}
